lumen_runtime/src/otp/erlang/tl_1.rs
// wasm32 proptest cannot be compiled at the same time as non-wasm32 proptest, so disable tests that
// use proptest completely for wasm32
//
// See https://github.com/rust-lang/cargo/issues/4866
#[cfg(all(not(target_arch = "wasm32"), test))]
mod test;

use std::convert::TryInto;

use liblumen_alloc::erts::exception;
use liblumen_alloc::erts::term::prelude::*;

use lumen_runtime_macros::native_implemented_function;

#[native_implemented_function(tl/1)]
pub fn native(list: Term) -> exception::Result {
    let cons: Boxed<Cons> = list.try_into()?;

    Ok(cons.tail)
}
